pyproject.toml
[project]
name = "hugging-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "diffusers",
    "huggingface-hub>=1.9.0",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
    "transformers>=5.5.0",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers" }
